Refactor JWT and refresh token models; update relationships and logout logic
