--- app/models/jwt_model.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-    
-class TokenData(BaseModel):
-    user_id: int
-
