+++ app/models/user_model.py
@@ -1,5 +1,8 @@
-from typing import Annotated
-from sqlmodel import Field, SQLModel
+from sqlmodel import Field, SQLModel, Relationship
+from typing import TYPE_CHECKING, List
+
+if TYPE_CHECKING:
+    from app.models.refresh_token_model import RefreshToken
 
 class User(SQLModel, table=True):
     __tablename__ = "users"
@@ -10,3 +13,5 @@ class User(SQLModel, table=True):
     
     hashed_password: str
     disabled: bool | None = None
+    
+    refresh_token: List["RefreshToken"] = Relationship(back_populates="user")
