+++ app/schemas/token_schema.py
@@ -1,4 +1,12 @@
 from pydantic import BaseModel
 
+class Token(BaseModel):
+    access_token: str
+    refresh_token: str
+    token_type: str
+    
+class TokenData(BaseModel):
+    user_id: int
+    
 class LogoutRequest(BaseModel):
     refresh_token: str
